+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This file is used to list changes made in each version of the openstack-metering
 
 ## 10.0.1
 * Support vmware hypervisor
+* Sync conf files with Juno
 
 ## 10.0.0
 * Upgrading to Juno