+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This file is used to list changes made in each version of the openstack-metering
 ## 10.0.1
 * Support vmware hypervisor
 * Sync conf files with Juno
+* Set os_region_name properly
 
 ## 10.0.0
 * Upgrading to Juno